refactor: 빠진 부분 수정

+++ domain/src/main/java/com/dragons/domain/payment/Payment.java
@@ -53,7 +53,7 @@ public class Payment extends BaseEntity {
     return payment;
   }
 
-  public static Payment withId(Long id, String holderName, int amount, String planType, String paymentType) {
+  public static Payment withId(Long id, String holderName, long amount, String planType, String paymentType) {
     Payment payment = new Payment();
     payment.setIdForTest(id);
     payment.holderName = holderName;
